requirements.txt
discord.py==1.7.2
pyodbc==4.0.32
python-dotenv==0.19.0
spotipy==2.19.0
twitchio==2.0.6
youtube-dl==2021.6.6